src/main/java/com/example/equityorderservice/controller/OrderController.java
package com.example.equityorderservice.controller;

import com.example.equityorderservice.model.Order;
import com.example.equityorderservice.service.OrderService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/orders")
public class OrderController {

    @Autowired
    private OrderService orderService;

    // Save a new order
    @PostMapping
    public Order createOrder(@RequestBody Order order) {
        return orderService.saveOrder(order);
    }

    // Update an existing order
    @PutMapping("/{id}")
    public Order updateOrder(@PathVariable Long id, @RequestBody Order orderDetails) {
        return orderService.updateOrder(id, orderDetails);
    }

    // Get all orders
    @GetMapping
    public List<Order> getOrders() {
        return orderService.getAllOrders();
    }
}
